+++ internal/app/app.go
@@ -75,7 +75,7 @@ func New() (*App, error) {
 	store := redisStore.New(cfg, rdb, "session", 86400*30)
 
 	// Create DB repositories
-	catsRepo, err := catsRepo.New(db)
+	catsRepo, err := catsRepo.New(db, nil)
 	if err != nil {
 		return nil, fmt.Errorf("couldn't create categories repo: %w", err)
 	}
